refactor: Consolidate sidebar components and add navigation in new directory for better organization
