+++ src/app/app/components/sidebar/index.ts
@@ -1,0 +1,11 @@
+export * from './sidebar'
+
+export * from './sidebar-navigation'
+export * from './sidebar-navigation-item'
+
+export * from './sidebar-search'
+export * from './sidebar-search-group'
+export * from './sidebar-search-movie'
+export * from './sidebar-search-person'
+export * from './sidebar-search-provider'
+export * from './sidebar-search-tv-show'
