+++ src/app/app/components/sidebar/sidebar-search-movie.tsx
