+++ src/app/app/components/sidebar/sidebar-search-provider.tsx
