+++ src/app/app/layout.tsx
@@ -3,9 +3,9 @@ import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
 
 import { Header } from './components/header'
-import { Sidebar } from './components/sidebar'
 import { ListsContextProvider } from '@/context/lists'
 import { AuthContextProvider } from '@/context/auth'
+import { Sidebar } from './components/sidebar'
 
 export default async function AppLayout({
   children,
